+++ flask/blueprints.py
@@ -13,6 +13,9 @@ from functools import update_wrapper
 
 from .helpers import _PackageBoundObject, _endpoint_from_view_func
 
+# a singleton sentinel value for parameter defaults
+_sentinel = object()
+
 
 class BlueprintSetupState(object):
     """Temporary holder object for registering a blueprint with the
@@ -90,6 +93,11 @@ class Blueprint(_PackageBoundObject):
     or other things on the main application.  See :ref:`blueprints` for more
     information.
 
+    .. versionchanged:: 1.1.0
+        Blueprints have a ``cli`` group to register nested CLI commands.
+        The ``cli_group`` parameter controls the name of the group under
+        the ``flask`` command.
+
     .. versionadded:: 0.7
     """
 
@@ -129,6 +137,7 @@ class Blueprint(_PackageBoundObject):
         subdomain=None,
         url_defaults=None,
         root_path=None,
+        cli_group=_sentinel,
     ):
         _PackageBoundObject.__init__(
             self, import_name, template_folder, root_path=root_path
@@ -142,6 +151,7 @@ class Blueprint(_PackageBoundObject):
         if url_defaults is None:
             url_defaults = {}
         self.url_values_defaults = url_defaults
+        self.cli_group = cli_group
 
     def record(self, func):
         """Registers a function that is called when the blueprint is
@@ -205,6 +215,17 @@ class Blueprint(_PackageBoundObject):
 
         for deferred in self.deferred_functions:
             deferred(state)
+
+        cli_resolved_group = options.get("cli_group", self.cli_group)
+
+        if cli_resolved_group is None:
+            app.cli.commands.update(self.cli.commands)
+        elif cli_resolved_group is _sentinel:
+            self.cli.name = self.name
+            app.cli.add_command(self.cli)
+        else:
+            self.cli.name = cli_resolved_group
+            app.cli.add_command(self.cli)
 
     def route(self, rule, **options):
         """Like :meth:`Flask.route` but for a blueprint.  The endpoint for the