+++ CHANGES.rst
@@ -56,6 +56,9 @@ Unreleased
     returning a string will produce a ``text/html`` response, returning
     a dict will call ``jsonify`` to produce a ``application/json``
     response. :pr:`3111`
+-   Blueprints have a ``cli`` Click group like ``app.cli``. CLI commands
+    registered with a blueprint will be available as a group under the
+    ``flask`` command. :issue:`1357`.
 
 .. _#2935: https://github.com/pallets/flask/issues/2935
 .. _#2957: https://github.com/pallets/flask/issues/2957